command-center/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
gitpython==3.1.43
python-docx==1.1.2
lxml==5.3.0
anthropic>=0.45.0
